refactor: rename AnalyticsAppObjC project to ObjCExample
