--- Examples/AnalyticsAppObjC/AnalyticsAppObjC/AppDelegate/AppDelegate.h
@@ -1,14 +1,0 @@
-//
-//  AppDelegate.h
-//  AnalyticsAppObjC
-//
-//  Created by Satheesh Kannan on 07/05/25.
-//
-
-#import <UIKit/UIKit.h>
-#import <RudderStackAnalytics/Analytics.h>
-@interface AppDelegate : UIResponder <UIApplicationDelegate>
-
-
-@end
-
