--- Examples/AnalyticsAppObjC/AnalyticsAppObjC/AppDelegate/SceneDelegate.h
@@ -1,15 +1,0 @@
-//
-//  SceneDelegate.h
-//  AnalyticsAppObjC
-//
-//  Created by Satheesh Kannan on 07/05/25.
-//
-
-#import <UIKit/UIKit.h>
-
-@interface SceneDelegate : UIResponder <UIWindowSceneDelegate>
-
-@property (strong, nonatomic) UIWindow * window;
-
-@end
-
