--- Examples/AnalyticsAppObjC/AnalyticsAppObjC/CustomLogger/CustomLogger.h
@@ -1,17 +1,0 @@
-//
-//  CustomLogger.h
-//  AnalyticsAppObjC
-//
-//  Created by Satheesh Kannan on 27/05/25.
-//
-
-#import <Foundation/Foundation.h>
-#import <Analytics/Analytics-Swift.h>
-
-NS_ASSUME_NONNULL_BEGIN
-
-@interface CustomLogger : NSObject<RSALogger>
-
-@end
-
-NS_ASSUME_NONNULL_END
