--- Examples/AnalyticsAppObjC/AnalyticsAppObjC/ObjCCompat/ObjCCompat.h
@@ -1,16 +1,0 @@
-//
-//  ObjCCompat.h
-//  AnalyticsAppObjC
-//
-//  Created by Satheesh Kannan on 29/05/25.
-//
-
-#import <Foundation/Foundation.h>
-
-NS_ASSUME_NONNULL_BEGIN
-
-@interface ObjCCompat : NSObject
-
-@end
-
-NS_ASSUME_NONNULL_END
