--- Examples/AnalyticsAppObjC/AnalyticsAppObjC/ViewControllers/ViewController.h
@@ -1,14 +1,0 @@
-//
-//  ViewController.h
-//  AnalyticsAppObjC
-//
-//  Created by Satheesh Kannan on 07/05/25.
-//
-
-#import <UIKit/UIKit.h>
-
-@interface ViewController : UIViewController
-
-
-@end
-
