+++ Examples/ObjCExample/ObjCExample/AppDelegate/AppDelegate.h
@@ -1,0 +1,14 @@
+//
+//  AppDelegate.h
+//  ObjCExample
+//
+//  Created by Satheesh Kannan on 07/05/25.
+//
+
+#import <UIKit/UIKit.h>
+#import <RudderStackAnalytics/RudderStackAnalytics-Swift.h>
+@interface AppDelegate : UIResponder <UIApplicationDelegate>
+
+
+@end
+
