+++ Examples/ObjCExample/ObjCExample/AppDelegate/SceneDelegate.h
@@ -1,0 +1,15 @@
+//
+//  SceneDelegate.h
+//  ObjCExample
+//
+//  Created by Satheesh Kannan on 07/05/25.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface SceneDelegate : UIResponder <UIWindowSceneDelegate>
+
+@property (strong, nonatomic) UIWindow * window;
+
+@end
+
