+++ Examples/ObjCExample/ObjCExample/Assets.xcassets/AccentColor.colorset/Contents.json
