+++ Examples/ObjCExample/ObjCExample/Assets.xcassets/AppIcon.appiconset/Contents.json
