+++ Examples/ObjCExample/ObjCExample/Assets.xcassets/Contents.json
