+++ Examples/ObjCExample/ObjCExample/CustomLogger/CustomLogger.h
@@ -1,0 +1,17 @@
+//
+//  CustomLogger.h
+//  ObjCExample
+//
+//  Created by Satheesh Kannan on 27/05/25.
+//
+
+#import <Foundation/Foundation.h>
+#import <RudderStackAnalytics/RudderStackAnalytics-Swift.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface CustomLogger : NSObject<RSALogger>
+
+@end
+
+NS_ASSUME_NONNULL_END
