+++ Examples/ObjCExample/ObjCExample/ObjCCompat/ObjCCompat.h
@@ -1,0 +1,16 @@
+//
+//  ObjCCompat.h
+//  ObjCExample
+//
+//  Created by Satheesh Kannan on 29/05/25.
+//
+
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface ObjCCompat : NSObject
+
+@end
+
+NS_ASSUME_NONNULL_END
