+++ Examples/ObjCExample/ObjCExample/Storyboards/Base.lproj/LaunchScreen.storyboard
