+++ Examples/ObjCExample/ObjCExample/main.m
@@ -1,0 +1,18 @@
+//
+//  main.m
+//  ObjCExample
+//
+//  Created by Satheesh Kannan on 07/05/25.
+//
+
+#import <UIKit/UIKit.h>
+#import "AppDelegate.h"
+
+int main(int argc, char * argv[]) {
+    NSString * appDelegateClassName;
+    @autoreleasepool {
+        // Setup code that might create autoreleased objects goes here.
+        appDelegateClassName = NSStringFromClass([AppDelegate class]);
+    }
+    return UIApplicationMain(argc, argv, nil, appDelegateClassName);
+}
